Increase Covid Case Heatmap performance
