+++ client/src/Components/Utils/ComposeChart.js
@@ -48,11 +48,9 @@ const ComposeChart = (props) => {
       >
          <h4>{props.title}</h4>
 
-         <div style={{ width: "100%" }}>
-            <ResponsiveContainer width="100%" aspect={4.0 / 3.0}>
+         <div>
+            <ResponsiveContainer width="100%" aspect={10.0 / 3.0}>
                <ComposedChart
-                  width={500}
-                  height={400}
                   data={props.data}
                   margin={{
                      top: 10,
